+++ auto-matching-nuxt/src/components/organisms/TemplateForm.vue
@@ -1,0 +1,3 @@
+<template lang="pug">
+div
+</template>
