+++ auto-matching-nuxt/src/plugins/i18n-ja.js
@@ -5,7 +5,8 @@ export default {
         index: '',
         sites: 'サイト管理',
         'sites-id': 'サイト設定',
-        'posts-templates': 'テンプレート管理',
+        'posts-templates': '投稿テンプレート管理',
+        'posts-templates-id': '投稿テンプレート修正',
         'posts-auto': '定期実行設定',
         'posts-manual': 'マニュアル投稿'
       }
