+++ auto-matching-nuxt/src/types/template.d.ts
@@ -1,0 +1,9 @@
+interface Template {
+  id: number
+  title: string
+  body: string
+  createdAt: Date
+  updatedAt: Date
+}
+
+export default Template
